FGDs: fix some issues with items, and simplify
While looking into adding the upcoming custom key item to the FGD files,
I noticed some issues with the existing item definitions:

    - Some of the items (at least item_key* and item_sigil) were missing
      some of the fields and/or spawnflags which should be common to all
      items.

    - The code defining the items was generally a lot longer and more
      complicated than it needed to be, with a lot of duplication.

This commit fixes and improves matters by doing the following:

    - It replaces the old "Item" base class with two new base classes,
      "RespawnableItem" and "NonRespawnableItem", which between them
      define all of the common spawnflags and fields.

    - It adds new base classes named "Artifact", "Armor", and "Key",
      which allow us to simplify the definitions of those types of
      items.

    - It updates all of the item definitions to make sure that every
      definition is derived from one of the new base classes.

This should ensure that no items are missing spawnflags or fields that
other items aren't.  It also greatly simplifies the definitions for the
items, which now take up about half the number of lines that they used
to.