TB FGD: display all item_sigil models
While working on the previous commits, I noticed that the Trenchbroom
FGD would only ever display the first of the four models for item_sigil,
regardless of which spawnflag was set.  This commit makes it so that the
second, third, or fourth model will be displayed if the relevant
spawnflag is set.