FGDs: remove the sounds field from trigger_usekey
When testing the FGD files in Trenchbroom, I noticed that the "sounds"
field was displayed for trigger_usekey.  trigger_usekey doesn't support
this field.

The reason trigger_usekey ended up with "sounds" displayed was because
the FGD file based trigger_usekey on the "Trigger" base class, which
includes "sounds" in its definition.  To fix the issue, this commit
changes trigger_usekey so that it is no longer based on "Trigger"
(though it is still based on the base classes that "Trigger" is based
on).  The only valid field that would be lost as a result of this change
is "delay", therefore this commit also manually adds "delay" to the
trigger_usekey definition.