Tweak the existing reset_items check
This commit modifies the existing check for the worldspawn "reset_items"
field so that it will work with the new "reset_items 2" functionality
which is going to be added.  For clarity, it also moves the check from
PutClientInServer() to DecodeLevelParms() and merges it with the
existing check for serverflags and start.bsp.