Fix func_trains getting stuck in some id1 maps
Previously, some func_train entities in some of the stock id1 maps (at
least e2m7, e3m3, and e3m6) would end up getting "stuck" after their
first move if the map was played with progs_dump.  This is a case where
the behavior of progs_dump is not quite backwards-compatible with the
behavior of the id1 progs.dat.  This commit adds a workaround for the
issue: please see the comment in the func_train spawn function for
details.

Note that this commit only attempts to work around the issue for the
stock id1 maps, because it's nice to keep those maps functional for
playtesting purposes.  This commit makes no attempt to fix any other
vanilla maps (i.e. maps not made by Id Software) which might exhibit the
same issue, because the progs.dat is detecting maps by name, and it
would be impractical to identify and add detection for every vanilla map
which is affected.