Detect whether one of the id1 maps is being played
This commit makes it so that the known_release global will be set to the
value KNOWN_RELEASE_ID1 if one of the stock id1 maps is being played.
A follow-up commit will use this to work around a compatibility issue.

This commit also tweaks the developer message printed by
DetectKnownRelease() just to make it a bit clearer.