boss2: fix kill count and target bugs
Previously, when a monster_boss2 was killed, it would count as two kills
if it died without being gibbed (allowing > 100% kills), and it would
activate its targets twice instead of once (whether it got gibbed or
not).

These are the same issues that my other recent commit fixed for
monster_oldone2, and the cause is the same: the code shouldn't increment
killed_monsters or call SUB_UseTargets() because the Killed() function
(in combat.qc) will already have taken care of those things.

For monster_boss2, commit 70a26be already fixed the kill count bug in
the case where the boss gets gibbed, but the bug would still happen if
it didn't get gibbed because killed_monsters was still incremented in
boss2_death10().

This commit fixes the issues by removing the unnecessary lines from
GibBoss2() and boss2_death10().