boss2: fix a "NULL function" Host_Error
The existing code for monster_boss2 had a critical bug which would occur
in the following situation:

    1. An entity other than the boss' current enemy deals damage to it
       via T_Damage().

    2. The boss doesn't die or go into its pain animation.

This is likely to happen if the boss gets hit by another monster's
attack, or in Coop, where multiple players are attacking the boss.

What happens in this case is that T_Damage() calls FoundTarget(), which
calls HuntTarget(), which tries to make the boss "get mad" at its new
attacker.  HuntTarget() does this by scheduling the function stored in
self.th_run as the next self.think function.  The problem was that
monster_boss2 didn't have anything assigned to self.th_run, so when the
engine subsequently tried to call self.think(), the engine would quit to
console with a "NULL function" Host_Error.

This commit fixes the problem by setting self.th_run to boss2_missile1,
which is the first frame of the boss' attack animation.

Note that this wasn't an issue for the original monster_boss because it
couldn't take damage via T_Damage().