boss2: disable pain animation until fully risen
Previously, if the player shot a monster_boss2 during its
rising-out-of-the-lava animation, it would go into its pain animation,
in which it is already fully risen.  That looked weird, because the boss
would seem to spring up to its full height instantaneously.

To prevent that from happening, this commit makes it so that
monster_boss2 can't go into its pain animation for a couple of seconds
after being awakened, which is enough time for it to finish its rising
animation.