FGDs: fix oldone/oldone2 bounding box size
The bounding box which the FGD files specified for monster_oldone was
much smaller than her in-game bounding box.  I've noticed the same bug
in other editor configuration files in the past.  I assume the reason
for this is that the QUAKED comment for monster_oldone always specified
an incorrect bounding box size, and that mistake just got propagated.

This commit fixes the bounding box sizes for both monster_oldone and
monster_oldone2 in both of the FGD files, and also fixes the QUAKED
comments.