Fix the falling obituary (id1 bug)
The "[player] fell to his death" obituary message was always unreliable:
it wouldn't necessarily be displayed if the player died by falling, and
it could be erroneously displayed instead of the generic "[player] died"
message in situations where the player didn't die by falling.

ClientObituary() prints the "fell to his death" message if the player
dies with their "deathtype" field set to the string "falling".  The
problem was that "deathtype" was neither set nor cleared at quite the
right places in the code.  This commit fixes that.

A side note, which is also relevant to the following commits:

The way Quake's obituary messages are implemented is dubious.  The
ClientObituary() function tries to guess how the player died based on
the state of the game at the time of the player's death, and then prints
what it thinks is the cause of death.  This guess-work is messy, hence
why ClientObituary() is such a long function; and it's also prone to
error, hence the existence of bugs like the ones fixed by this commit
and the commits that follow.

A more thorough solution would be to overhaul the whole obituary system,
and make it so that (for example) the desired obituary message is
specified as a parameter to T_Damage() at the time damage is dealt,
either as a string or, for the more complex obituaries, as a function.
However, that sort of overhaul would be a major operation, so in this
commit and the commits that follow, I've aimed to fix the buggy obituary
messages in whatever way seems simplest.