Fix precedence of the liquid obituaries (id1 bug)
This commit fixes a minor bug from the original game, which is that the
drowning/slime/lava obituary messages would always be displayed in
preference to the generic "[player] died" obituary message if the player
died while in water/slime/lava, even if the player didn't actually die
because of drowning/slime/lava.

To fix the issue, this commit makes it so that the code in client.qc
explicitly selects the drowning/slime/lava obituaries by setting the
player's "deathtype" field, like how the falling obituary works.