Fix the func_togglewall obituary (SoA bug)
Previously, if the player died by touching a func_togglewall, they would
end up with the "[player] was squished" obituary message which is
usually used for plats, trains, etc.  But that message doesn't make
sense for a func_togglewall because it isn't a moving object.

The reason func_togglewall ended up triggering the "squished" obituary
is because it's SOLID_BSP, and ClientObituary() effectively uses the
"squished" obituary as a default obituary for SOLID_BSP entities.

This commit makes it so that the "squished" obituary will not be used
for func_togglewall, with the result that it will give the player the
generic "[player] died" obituary message instead.  This seems correct
because a func_togglewall could represent different things in different
maps, so we can't really be more specific than that.