Fix the func_laser obituary
Previously, if a player was killed by a func_laser that had the
LASER_SOLID spawnflag set (like the ones in pd_lasers), they would get
the "[player] was squished" obituary message.  This happened for the
same reason as the func_togglewall issue which I fixed in the previous
commit: a func_laser with LASER_SOLID is SOLID_BSP, and ClientObituary()
defaults to the "[player] was squished" message when the attacker is
SOLID_BSP.

In Rubicon 2, this wasn't an issue because the ClientObituary() function
included a special obituary message just for func_laser, and it checked
whether it should print that message before doing the SOLID_BSP check.

The obvious solution is to make progs_dump do the same thing.  However,
the message used for func_laser in Rubicon 2 was a pretty obscure joke
("[player] looked into laser with remaining eye"), and I'm not sure
whether most players would understand it.  So, for this commit I just
made up a new message for func_laser.

dumptruck_ds, if you'd prefer to have a different message for
func_laser, please just change the string to whatever you prefer!