Use the "blew up" obituary for play_* explosions
Previously, a player killed by a play_explosion or a play_tbabyexplode
would just get the generic "[player] died" obituary message.  This
commit makes it so that the player will get the "[player] blew up"
message, the same as if they were killed by a misc_explobox or
misc_explobox2.