Allow custom obituaries via the deathtype field
Although this commit is a new feature, it's also a bug fix, and that's
my reason for making it.

The manual and the FGD files indicate that the mapper should be able to
set custom obituary messages for the func_rotate_* entities by setting
their "deathtype" field.  However, previously this didn't work.

The reason why "deathtype" could be set for the func_rotate_* entities
in Scourge of Armagon (where the func_rotate_* stuff originates) is
because the SoA progs.dat allowed "deathtype" to be set on more-or-less
anything that could kill the player.

This commit implements the same functionality for progs_dump.  It also
goes a step further by allowing the mapper to set "deathtype" on either
an "inflictor" entity or an "attacker" entity (in SoA, it only worked
when set on an attacker).  The inflictor and attacker are often one and
the same entity, but there are some cases where they are not: for
example, if a player gets crushed by a func_movewall which is controlled
by a func_rotate_entity, the func_movewall is the inflictor and the
func_rotate_entity is the attacker.  I didn't think the mapper should
have to care about this distinction.

In order to make this work, the Killed() and ClientObituary() functions
have been changed so that they now take the inflictor entity as
a parameter in addition to the parameters they already had.  Killed() is
only ever called by T_Damage(), and ClientObituary() is only ever called
by Killed(), so this change doesn't affect other parts of the code.