Fix an error check for trigger_usekey
The spawnflag check in the trigger_usekey spawn function had the same
issue that the similar checks in the weapon_shotgun and func_bob spawn
functions used to have (they were fixed in commits 0eaf6eb and a79d952
respectively): the code checked whether no spawnflags were set, whereas
to behave correctly, it needs to check whether specific spawnflags
aren't set.  This commit fixes the test and updates the wording of the
objerror message accordingly.

This commit also removes the "\n" from the end of the objerror message
(it's not needed when calling objerror), and also gets rid of the line
which removed self (objerror already takes care of that).