Use named constants for the worldtype values
The original QuakeC code always compared the value of world.worldtype
against magic numbers: 0 for "medieval", 1 for "metal", and 2 for
"base".  This commit adds constants named WORLDTYPE_MEDIEVAL,
WORLDTYPE_METAL, and WORLDTYPE_BASE, and updates the rest of the code to
use them instead of the numbers.  This just makes the code a bit easier
to read.