Add centerprint2() through centerprint7()
(This is part of the prep work for implementing item_key_custom.)

Please see the comment in the code for an explanation of how and why
these function definitions work.  I'm going to need to use one of these
functions in a follow-up commit, and considering that I need to add one
of them, I thought that I might as well add the whole set so they'll be
available for future use.