Consolidate code involving key names
(This is part of the prep work for implementing item_key_custom.)

In the QuakeC code for the original game, the logic to work out the name
of a key based on the current worldtype (e.g. "silver key", "silver
runekey", or "silver keycard") was duplicated in different places,
appearing not only in the code for item_key1 and item_key2 but also in
the code for func_door.  In this mod, the same logic was duplicated
again in the code for trigger_usekey.

This commit does two things to simplify the code:

- It adds two new functions, SilverKeyName() and GoldKeyName(), each of
  which returns the name that should be used for their respective key,
  and updates other parts of the code to call these functions instead of
  duplicating the same logic.

- It tweaks the code for func_door and trigger_usekey so that these
  entities now store the name of their key in their self.netname field
  (which these entities previously didn't use).  This simplifies the
  logic in the door_touch() and keytrigger_use() functions.