Consolidate code for item_key1 and item_key2
(This is part of the prep work for implementing item_key_custom.)

There were some common lines of code which were duplicated in the spawn
functions for item_key1 and item_key2.  This commit moves the common
code into a new function, key_start(), and updates the two spawn
functions so that they call this new function.