Display a message when a key is used
(This is part of the prep work for implementing item_key_custom.)

The original game did not display a message when the player used a key
to unlock something: the key icon would simply disappear from the
player's status bar.  A follow-up commit is going to implement a new
type of key item, item_key_custom, which unfortunately will not have
a status bar icon (because the status bar simply isn't very
customizable, at least in traditional Quake engines).  So, in order to
make it clear to the player that they have actually used a key, this
commit makes it so that an unobtrusive message is displayed at the
top-left corner of the screen when a key is used.