Allow custom noise3/4 for func_door/trigger_usekey
The new item_key_custom entity already allows the mapper to override the
regular key pickup sound by setting the entity's "noise" field; the
regular key pickup sound has always been worldtype-specific.  However,
func_door and trigger_usekey also have worldtype-specific sound effects
which indicate to the player that a key is required or that a key has
been used.  It would be nice if the mapper could override those, too.

The paths for the "key required" and "key used" sound files are
respectively stored in the "noise3" and "noise4" fields of the func_door
or trigger_usekey.  This commit makes it so that the mapper can override
the default sound files by setting these fields.