Allow custom noise1/2 for func_door
Considering that the mapper can now specify custom sound files for the
"noise3" and "noise4" fields of func_door, it only makes sense to allow
custom sound files for "noise1" and "noise2" as well; "noise1" is the
"stop moving" sound effect, and "noise2" is the "move" sound effect.
This commit allows the mapper to set these fields.  If set, these fields
take precedence over the vanilla "sounds" field.