Use noise1/2 instead of 3/4 for trigger_usekey
On reflection, it seemed silly that the mapper-settable custom sound
fields for trigger_usekey were "noise3" and "noise4", considering that
trigger_usekey doesn't use "noise1" or "noise2" for anything.  Even
though the code uses "noise3" and "noise4" internally (because the code
is shared with func_door), we don't have to expose that fact to the
mapper.

This commit makes it so that the mapper-settable fields are "noise1" for
the "key required" sound and "noise2" for the "key used" sound.