Fix the func_movewall obituary (SoA bug)
A func_movewall is the solid but invisible component of a rotating
entity setup.  They're capable of killing the player in two different
ways:

1.  By crushing the player.

2.  By dealing damage to the player on contact, which only happens if
    the MOVEWALL_TOUCH spawnflag is set.

Logically, the second kind of func_movewall should give the player the
generic "[player] died" obituary message, because we don't know what the
mapper wants it to represent: it could be a giant fan, a grinder, or
something else entirely.  However, the first kind of func_movewall
should give the player the usual "[player] was squished" obituary,
because that's what we expect when the player gets crushed.

Previously, func_movewall always gave the player the "[player] died"
obituary.  This commit makes it give either the "died" or "was squished"
obituary depending on circumstances.