Gracefully handle untargeted sleeping zombies
Previously, if a map contained a sleeping zombie with either no
targetname or an unmatched targetname, it would be impossible for the
player to get 100% kills because a sleeping zombie can't be killed until
it has been woken up.

The zombie_start() function already contained a test which checked
whether a sleeping zombie had a targetname, however it didn't check
whether the targetname was matched, and didn't attempt to fix the kill
count.

This commit replaces the old test with a new test which is very similar
to what I implemented in commit d23382f for trigger-spawned monsters.
The new test has the following improvements:

- It considers self.targetname2/3/4 as well as self.targetname.

- It checks that one of the targetnames matches the target, target2/3/4,
  or pain_target of another entity.

- It corrects the kill count when removing the sleeping zombie in the
  case where the zombie doesn't have a matched targetname.