Reinstate new spawnflags for misc_teleporttrain
Commit e58d385 replaced the code for misc_teleporttrain with all-new,
fixed code, which is very much appreciated (thanks ZungryWare!), however
this had the side-effect of removing the call to SUB_Inhibit() which was
at the top of the misc_teleporttrain spawn function.  SUB_Inhibit()
implements the new global spawnflags for this mod (NOT_IN_COOP etc.),
so, this commit puts that call back in.