Fix the recent shambler projectile bug fix
Commit a899a57 fixed the usual case of the vanilla Quake bug whereby
a projectile can sometimes pass straight through a shambler, however it
wasn't quite a complete or robust solution.

In particular, it didn't handle the case where the shambler switches
into a different animation sequence (other than its pain animation)
after spawning the lightning arc in sham_magic3() but before removing
the arc in sham_magic6().  That can happen e.g. if the shambler gets
damaged by another monster and "gets angry" at it (without also going
into its pain animation).  In that case, HuntTarget() puts the shambler
back into its self.th_run() sequence, and after the lightning arc
removes itself, the shambler's self.owner will be a dangling reference,
which is of course what caused the original bug.

It looks like the new code in sham_pain() could interact badly with
this, because sham_pain() only tests whether self.owner isn't world
before changing self.owner.frame; however, a dangling reference won't be
world, and so will pass this test.

This commit does the following:

- To eliminate the original vanilla Quake bug once and for all,
  a shambler now saves its lightning arc in self.trigger_field instead
  of self.owner.  The reason the collision bug existed in the first
  place is because the engine treats self.owner specially (an entity
  won't collide with the entity that is its owner), but the shambler
  doesn't want or need that special behavior.

- The shambler's lightning arc now keeps track of the shambler that
  spawned it (by making the shambler the arc's owner, which should be
  safe).  In the case where the lightning arc removes itself, it will
  tidy up the shambler's trigger_field first, which should prevent
  a dangling reference.

- The test in sham_pain() now has an additional sanity-check before the
  shambler attempts to change the lightning arc's frame.

To be honest, the code for the lightning arc has always been kind of
a mess and probably deserves rewriting from scratch.  But it's been
years since I looked at this code, and I'd rather not risk b0rking it so
close to the map jam, so I'm leaving it at this for now.